// teams/people/organizations-management/base/src/domains/agents-v1/core/entity.schema.types.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface IAgentV1Dto {
  /**
   * The unique identifier of an agent as a hexadecimal string of 28 characters.
   */
  id: string;
  /**
   * The nickname of the agent. It must have at least 4 characters.
   */
  nickname: string;
  /**
   * The email of the agent. An organization agent can have the same e-mail of its owner only if its owner is an individual agent. Two individual agents cannot have the same e-mail.
   */
  email: string;
  /**
   * The type of the agent. It can be individual or organization.
   */
  type: "INDIVIDUAL" | "ORGANIZATION";
  /**
   * The date and time when the agent was created.
   */
  createdAt: string;
  /**
   * The date and time when the agent was last updated.
   */
  updatedAt: string;
}